Refactor: outsource form into component
